Add prompt to avoid accidental deletion of training logs.

+++ clean.sh
@@ -1,3 +1,8 @@
-rm *.out
-rm *.log
-rm *.err
+read -p "Are you sure you want to delete log files? (Remember to save the perplexity logs from training!) " -n 1 -r
+echo
+if [[ $REPLY =~ ^[Yy]$ ]]
+then
+	rm *.out
+	rm *.log
+	rm *.err
+fi
